WIP - Removing left bar. Removed everything except File menu.
